fix logInToFirebase - change to async function and make layout call with useEffect instead remove redundant NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET
