+++ .env
@@ -19,8 +19,6 @@ FIREBASE_STORAGE_BUCKET =
 FIREBASE_PRIVATE_KEY =
 FIREBASE_PROJECT_ID =
 
-NEXT_PUBLIC_FIREBASE_STORAGE_BUCKET = 
-
 FIREBASE_SERVICE_ACCOUNT = 
 
 # Firebase
