+++ src/app/dashboard/layout.tsx
@@ -1,12 +1,20 @@
 "use client";
-import logInToFirebase from "@/lib/tokenFetcher.client";
+import { logInToFirebase } from "@/lib/tokenFetcher.client";
+import { useEffect } from "react";
+import { usePathname } from "next/navigation";
+
 
 export default function DashboardLayout({
   children,
 }: {
   children: React.ReactNode;
-}) {  
-	logInToFirebase();
+}) {
+  const pathname = usePathname();
+
+  useEffect(() => {
+    logInToFirebase();
+  }, [pathname]);
+
   return (
     <>
       {children}
