+++ src/lib/tokenFetcher.client.ts
@@ -1,0 +1,32 @@
+"use client";
+
+import { useEffect } from "react";
+import { usePathname } from "next/navigation";
+import { signInWithCustomToken } from "firebase/auth";
+import { singletonFirebaseAuth } from "@/lib/singleton/firebaseAuth.client";
+
+export async function logInToFirebase() {
+  
+  async function fetchToken() {
+    const storedExpiry = localStorage.getItem("tokenExpiry");
+    if (storedExpiry && parseInt(storedExpiry) > Date.now()) return;
+    const res = await fetch("/api/get-token");
+    const data = await res.json();
+    if (data.token){
+      localStorage.setItem("token", data.token);
+      localStorage.setItem("tokenExpiry", data.expiresAt);
+    }
+    // console.log("token received:", data.token);
+
+  }
+  async function signInWithToken() {
+    const token = localStorage.getItem("token");
+    if (!token) return;
+    console.log("signing in to firebase with token:", "**HIDDEN**");
+    await signInWithCustomToken(singletonFirebaseAuth, token);
+  }
+  fetchToken();
+  signInWithToken();
+
+  return null;
+}
